+++ src/runtime/errors.go
@@ -1,4 +1,4 @@
-package common
+package runtime
 
 import "errors"
