+++ src/runtime/runtime_address_info_test.go
@@ -1,4 +1,4 @@
-package common
+package runtime
 
 import (
 	"net/netip"
