+++ src/runtime/runtime_session_test.go
@@ -1,4 +1,4 @@
-package common
+package runtime
 
 import (
 	"context"
